JavaScript operators
JavaScript operators can be classified into several categories:

Arithmetic Operators: These operators are used to perform arithmetic operations such as addition, subtraction, multiplication, division, and modulo (remainder). They allow you to manipulate numerical values in your code.

Assignment Operators: Assignment operators are used to assign values to variables. They include simple assignment (=) as well as compound assignment operators like +=, -=, *=, /=, and %=, which combine an arithmetic operation with assignment.

Comparison Operators: Comparison operators are used to compare two values and return a Boolean result (true or false) based on the comparison. Examples include ==, ===, !=, !==, >, <, >=, and <=.

Logical Operators: Logical operators are used to combine multiple conditions and perform logical operations such as AND (&&), OR (||), and NOT (!). They are commonly used in conditional statements and loops to control the flow of execution.

Unary Operators: Unary operators act on a single operand and are used to perform various operations such as negation (-), increment (++), decrement (--), type conversion (typeof), and logical negation (!).

Ternary Operator (Conditional Operator): The ternary operator (a ? b : c) is a compact way of writing conditional statements. It evaluates the condition (a) and returns one of two possible values (b or c) based on whether the condition is true or false.